Fix overly strict polar check in NauticalAngles that produced NaNs
The added test cases demonstrates that for some polar values, we'll end
up with a v2.getZ() that is outside of the domain for asin due to
floating point errors in the LSBs (i.e., -1.0000000000000002). In that
case, we need to substitute the limit of asin(x) as x goes to -1 or 1
to avoid the NaNs (i.e., -PI / 2 or PI / 2).