Ensure that the Cesium Viewer can render when offline
If the terrainProvider does not have tiles available, then the globe
will completely fail to render. This change initializes the globe with
an ellipsoid (WGS84) and switches over to a network terrain server if
the terrain server is available at page load. While it is possible to
switch the terrain on the fly, the extra complexity of adding a UI
element to do that is difficult to justify against reloading the page.